+++ server/tests/backends/mongo_translator/test_mongo_translator_steps.py
@@ -4,6 +4,7 @@ import uuid
 from io import StringIO
 
 import docker
+import geopandas as gpd
 import pandas as pd
 import pytest
 from pandas.api.types import is_bool_dtype, is_datetime64_any_dtype, is_numeric_dtype
@@ -47,6 +48,14 @@ def mongo_database(mongo_server_port):
     return client["tests"]
 
 
+def _serialize_geo_df(gdf: pd.DataFrame) -> pd.DataFrame:
+    """Converting the GeoSeries in shapely format to GeoJSON features"""
+    gdf = gpd.GeoDataFrame(gdf)
+    geo_series = gdf.geometry
+    gdf[geo_series.name] = json.loads(geo_series.to_json())["features"]
+    return pd.DataFrame(gdf)
+
+
 def _sanitize_df(df: pd.DataFrame) -> pd.DataFrame:
     non_numeric_cols = [col for col, dtype in df.dtypes.items() if not is_numeric_dtype(dtype)]
     # Ensure we have None for non-numeric columns such as datetimes or
@@ -69,8 +78,18 @@ def _sanitize_df(df: pd.DataFrame) -> pd.DataFrame:
     return df[sorted(df.columns)]  # order of columns may be different between pandas and mongo
 
 
+def _load_df(spec: dict) -> pd.DataFrame:
+    return (
+        pd.DataFrame(gpd.read_file(json.dumps(spec["data"])))
+        if spec.get("schema") == "geojson"
+        else pd.read_json(StringIO(json.dumps(spec)), orient="table")
+    )
+
+
 def _sanitized_df_from_pandas_table(df_spec: dict) -> pd.DataFrame:
-    df = _sanitize_df(pd.read_json(StringIO(json.dumps(df_spec)), orient="table"))
+    df = _sanitize_df(_load_df(df_spec))
+    if df_spec["schema"] == "geojson":
+        return df
     bool_cols = [f["name"] for f in df_spec["schema"]["fields"] if f["type"] == "boolean"]
     # By default, pandas converts null bools to False. This enforces
     # use of nullable booleans
@@ -80,19 +99,25 @@ def _sanitized_df_from_pandas_table(df_spec: dict) -> pd.DataFrame:
     return df
 
 
+def insert_data_to_mongo(mongo_database, collection_uid: str, schema: dict | str, df: pd.DataFrame):
+    if schema == "geojson":
+        df = _serialize_geo_df(df)
+    mongo_database[collection_uid].insert_many(df.to_dict(orient="records"))
+
+
 @pytest.mark.parametrize("case_id,case_spec_file_path", test_cases)
 def test_mongo_translator_pipeline(mongo_database, case_id, case_spec_file_path, available_variables):
     # insert in mongoDB
     collection_uid = uuid.uuid4().hex
     spec = get_spec_from_json_fixture(case_id, case_spec_file_path)
-    data = _sanitized_df_from_pandas_table(spec["input"]).to_dict(orient="records")
-    mongo_database[collection_uid].insert_many(data)
+    data = _sanitized_df_from_pandas_table(spec["input"])
+    insert_data_to_mongo(mongo_database, collection_uid, spec["input"]["schema"], data)
     if "other_inputs" in spec and (
         "join" in case_id or "append" in case_id
     ):  # needed for join & append steps tests as we need a != collection
         for collection_name, raw_df in spec["other_inputs"].items():
-            df = pd.read_json(StringIO(json.dumps(raw_df)), orient="table")
-            mongo_database[collection_name].insert_many(df.to_dict(orient="records"))
+            df = _load_df(raw_df)
+            insert_data_to_mongo(mongo_database, collection_name, raw_df["schema"], df)
 
     # create query
     steps = spec["step"]["pipeline"]