+++ server/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Mongo: `join` step adds also new columns to rows that do not match join conditions.   
+
 ## [0.56.0] - 2025-05-07
 
 ### Breaking